Add AudioMixerFloat
Combines Get and Set commands that existed previously, makes more sense to mirror the behaivour of Fungus Property commands
