--- Assets/Fungus/Scripts/Commands/AudioMixerGetFloat.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f6766649723944f48805b1d9a52d972e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
