+++ Assets/FungusExamples/AudioMixer/AudioDemoScene.unity
@@ -501,10 +501,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   version: 1
-  scrollPos: {x: 0, y: 0}
+  scrollPos: {x: -144.37125, y: 367.01898}
   variablesScrollPos: {x: 0, y: 0}
   variablesExpanded: 1
-  blockViewHeight: 521
+  blockViewHeight: 586
   zoom: 1
   scrollViewRect:
     serializedVersion: 2
@@ -876,6 +876,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   itemId: 12
   indentLevel: 0
+  getOrSet: 1
   mixer:
     audioMixerRef: {fileID: 0}
     audioMixerVal: {fileID: 24100000, guid: 6bb64aab09f3f734193dd0ac7e1f5df0, type: 2}
